+++ Sources/ViewModels/ALKConversationListViewModel.swift
@@ -237,26 +237,6 @@ final public class ALKConversationListViewModel: NSObject, ALKConversationListVi
         }
     }
 
-    func titleFor(contactId: String?, channelId: NSNumber?) -> String {
-        var title = ""
-        if let key = channelId,
-            let alChannel = alChannelService.getChannelByKey(key),
-            let name = alChannel.name {
-            title = name
-        } else if let key = contactId,
-            let alContact = alContactService.loadContact(byKey: "userId", value: key),
-            let name = alContact.getDisplayName() {
-            title = name
-        }
-
-        let noName = localizedString(
-            forKey: "NoNameMessage",
-            withDefaultValue: SystemMessage.NoData.NoName,
-            fileName: localizationFileName)
-        title = title.isEmpty ? noName : title
-        return title
-    }
-
     func conversationViewModelOf(
         type conversationViewModelType: ALKConversationViewModel.Type,
         contactId: String?,